requirements.txt
Django==3.2.8
python-dotenv==0.19.1
razorpay==1.2.0
django-widget-tweaks==1.4.8
Pillow==8.3.2
